may fix ctcloss nan

+++ demo.py
@@ -44,7 +44,7 @@ preds = model(image)
 _, preds = preds.max(2)
 preds = preds.transpose(1, 0).contiguous().view(-1)
 
-preds_size = Variable(torch.IntTensor([preds.size(0)]))
+preds_size = Variable(torch.LongTensor([preds.size(0)]))
 raw_pred = converter.decode(preds.data, preds_size.data, raw=True)
 sim_pred = converter.decode(preds.data, preds_size.data, raw=False)
 print('%-20s => %-20s' % (raw_pred, sim_pred))
